+++ src/Verify.NUnit.Tests/Snippets/ParametersSample.cs
@@ -12,7 +12,7 @@ public class ParametersSample
     }
 
     [TestCase((float) 1.1)]
-    public async Task Decimal(float arg)
+    public async Task Float(float arg)
     {
         await Verifier.Verify(arg)
             .UseParameters(arg);
